Move StepMatchSearch into own file
