+++ lib/cucumber/filters/activate_steps.rb
@@ -5,16 +5,16 @@ require 'cucumber/errors'
 
 module Cucumber
   module Filters
-    class ActivateSteps < Core::Filter.new(:step_definitions, :configuration)
+    class ActivateSteps < Core::Filter.new(:step_match_search, :configuration)
 
       def test_case(test_case)
-        CaseFilter.new(test_case, step_definitions, configuration).test_case.describe_to receiver
+        CaseFilter.new(test_case, step_match_search, configuration).test_case.describe_to receiver
       end
 
       class CaseFilter
-        def initialize(test_case, step_definitions, configuration)
+        def initialize(test_case, step_match_search, configuration)
           @original_test_case = test_case
-          @step_definitions = step_definitions
+          @step_match_search = step_match_search
           @configuration = configuration
         end
 
@@ -33,12 +33,12 @@ module Cucumber
         end
 
         def find_match(test_step)
-          StepMatchSearch.new(@step_definitions, @configuration, test_step).result
+          FindMatch.new(@step_match_search, @configuration, test_step).result
         end
 
-        class StepMatchSearch
-          def initialize(step_match_library, configuration, test_step)
-            @step_match_library, @configuration, @test_step = step_match_library, configuration, test_step
+        class FindMatch
+          def initialize(step_match_search, configuration, test_step)
+            @step_match_search, @configuration, @test_step = step_match_search, configuration, test_step
           end
 
           def result
@@ -50,15 +50,15 @@ module Cucumber
 
           private
 
-          attr_reader :step_match_library, :configuration, :test_step
-          private :step_match_library, :configuration, :test_step
+          attr_reader :step_match_search, :configuration, :test_step
+          private :step_match_search, :configuration, :test_step
 
           def match
             matches.first
           end
 
           def matches
-            step_match_library.step_matches(test_step.name)
+            step_match_search.call(test_step.name)
           end
         end
       end
